Refactor kanva: KanvaSink, PaintOrder, prelude, wrap_comments
- Replace PaintSink with KanvaSink in KanvaBuilder
- Add PaintOrder to KanvaPath; render respects draw order
- Consolidate public types into prelude
- Extract kanva_svg convert functions into convert.rs
- Rewrite kanva/src/lib.rs tests for current architecture
- Enable wrap_comments in rustfmt.toml
