+++ crates/kanva/src/lib.rs
@@ -14,11 +14,9 @@ pub mod node;
 pub mod sink;
 
 pub use imaging;
-pub use modifiers::{
-    GroupModEntry, GroupMods, PathModEntry, PathMods,
-};
-pub use node::*;
-pub use sink::{GlyphRun, KanvaSink};
+
+use modifiers::{GroupModEntry, GroupMods, PathModEntry, PathMods};
+use node::*;
 
 pub mod prelude {
     pub use crate::Kanva;
@@ -28,27 +26,29 @@ pub mod prelude {
     };
     pub use crate::node::{
         Command, Group, GroupRange, KanvaClip, KanvaFill, KanvaPath,
-        KanvaStroke, NodeIndex,
+        KanvaStroke, NodeIndex, PaintOrder,
     };
     pub use crate::sink::{GlyphRun, KanvaSink};
 }
 
 /// A baked 2D graphics scene graph.
 ///
-/// Stores paths, fills, strokes, and groups in flat parallel buffers indexed
-/// by a [`Command`] buffer that encodes draw order. Groups may carry a
-/// transform that is accumulated onto child paths at render time without
-/// modifying stored data.
+/// Stores paths, fills, strokes, and groups in flat parallel buffers
+/// indexed by a [`Command`] buffer that encodes draw order. Groups
+/// may carry a transform that is accumulated onto child paths at
+/// render time without modifying stored data.
 ///
-/// Primary data is write-once at build time. Per-field overrides are stored
-/// in [`PathMods`] and [`GroupMods`] and reset via [`Self::clear_mods`].
-/// Each field map is keyed by path or group index; absent = keep stored value.
-/// For optional-target fields (`fill`, `stroke`, `clip`), `None` clears the field.
+/// Primary data is write-once at build time. Per-field overrides are
+/// stored in [`PathMods`] and [`GroupMods`] and reset via
+/// [`Self::clear_mods`]. Each field map is keyed by path or group
+/// index; absent = keep stored value. For optional-target fields
+/// (`fill`, `stroke`, `clip`), `None` clears the field.
 #[derive(Default, Debug, Clone)]
 pub struct Kanva {
     commands: Vec<Command>,
     groups: Vec<Group>,
-    /// Parallel to `groups`: command index range (push, pop) for each group.
+    /// Parallel to `groups`: command index range (push, pop) for
+    /// each group.
     group_cmds: Vec<GroupRange>,
     paths: Vec<KanvaPath>,
     fills: Vec<KanvaFill>,
@@ -65,7 +65,8 @@ impl Kanva {
         Self::default()
     }
 
-    /// Returns `true` if the kanva has no commands (nothing to render).
+    /// Returns `true` if the kanva has no commands (nothing to
+    /// render).
     pub fn is_empty(&self) -> bool {
         self.commands.is_empty()
     }
@@ -79,7 +80,8 @@ impl Kanva {
 
     /// Look up a labeled group and return its raw index directly.
     ///
-    /// Returns `None` if the label does not exist or resolves to a path.
+    /// Returns `None` if the label does not exist or resolves to a
+    /// path.
     pub fn query_group(&self, label: &str) -> Option<usize> {
         match self.index.get(label).copied()? {
             NodeIndex::Group(i) => Some(i),
@@ -89,7 +91,8 @@ impl Kanva {
 
     /// Look up a labeled path and return its raw index directly.
     ///
-    /// Returns `None` if the label does not exist or resolves to a group.
+    /// Returns `None` if the label does not exist or resolves to a
+    /// group.
     pub fn query_path(&self, label: &str) -> Option<usize> {
         match self.index.get(label).copied()? {
             NodeIndex::Path(i) => Some(i),
@@ -97,17 +100,20 @@ impl Kanva {
         }
     }
 
-    /// Returns the [`KanvaPath`] at `idx`, or `None` if out of bounds.
+    /// Returns the [`KanvaPath`] at `idx`, or `None` if out of
+    /// bounds.
     pub fn get_path(&self, idx: usize) -> Option<&KanvaPath> {
         self.paths.get(idx)
     }
 
-    /// Returns the [`KanvaFill`] at `idx`, or `None` if out of bounds.
+    /// Returns the [`KanvaFill`] at `idx`, or `None` if out of
+    /// bounds.
     pub fn get_fill(&self, idx: usize) -> Option<&KanvaFill> {
         self.fills.get(idx)
     }
 
-    /// Returns the [`KanvaStroke`] at `idx`, or `None` if out of bounds.
+    /// Returns the [`KanvaStroke`] at `idx`, or `None` if out of
+    /// bounds.
     pub fn get_stroke(&self, idx: usize) -> Option<&KanvaStroke> {
         self.strokes.get(idx)
     }
@@ -127,23 +133,45 @@ impl Kanva {
         &self.paths
     }
 
-    /// Returns the contiguous range of path indices directly owned by this group.
+    /// Returns the contiguous range of path indices directly owned by
+    /// this group.
     ///
-    /// Scans the group's inner commands (PushGroup/PopGroup excluded) for the
-    /// first and last [`Command::DrawPath`] indices and returns `first..last + 1`.
-    /// Returns `None` if the group index is out of bounds or the group has no paths.
+    /// Scans the group's inner commands (PushGroup/PopGroup excluded)
+    /// for the first and last [`Command::DrawPath`] indices and
+    /// returns `first..last + 1`. Returns `None` if the group
+    /// index is out of bounds or the group has no paths.
     /// Use [`Self::get_path`] to access individual paths by index.
     ///
     /// ```
-    /// use kanva::prelude::*;
     /// use kanva::imaging::kurbo::{Affine, BezPath};
     /// use kanva::imaging::peniko::{Brush, Fill};
+    /// use kanva::prelude::*;
     ///
     /// let mut builder = KanvaBuilder::new();
     /// let brush = Brush::default();
     /// builder.push_group(Group::default());
-    /// builder.draw_path(BezPath::new(), Affine::IDENTITY, Some(KanvaFill { rule: Fill::NonZero, brush: brush.clone(), ..Default::default() }), None);
-    /// builder.draw_path(BezPath::new(), Affine::IDENTITY, Some(KanvaFill { rule: Fill::NonZero, brush: brush.clone(), ..Default::default() }), None);
+    /// builder.draw_path(
+    ///     BezPath::new(),
+    ///     Affine::IDENTITY,
+    ///     Some(KanvaFill {
+    ///         rule: Fill::NonZero,
+    ///         brush: brush.clone(),
+    ///         ..Default::default()
+    ///     }),
+    ///     None,
+    ///     Default::default(),
+    /// );
+    /// builder.draw_path(
+    ///     BezPath::new(),
+    ///     Affine::IDENTITY,
+    ///     Some(KanvaFill {
+    ///         rule: Fill::NonZero,
+    ///         brush: brush.clone(),
+    ///         ..Default::default()
+    ///     }),
+    ///     None,
+    ///     Default::default(),
+    /// );
     /// builder.pop_group();
     /// let kanva = builder.build();
     ///
@@ -172,10 +200,11 @@ impl Kanva {
         Some(first..last + 1)
     }
 
-    /// Returns the contiguous range of path indices that appear in the command
-    /// stream between two labeled marker groups (exclusive of the groups
-    /// themselves). Useful when paths are placed between `<start>` and `<end>`
-    /// marker boxes rather than inside a single labeled group.
+    /// Returns the contiguous range of path indices that appear in
+    /// the command stream between two labeled marker groups
+    /// (exclusive of the groups themselves). Useful when paths
+    /// are placed between `<start>` and `<end>` marker boxes
+    /// rather than inside a single labeled group.
     pub fn get_paths_between_groups(
         &self,
         start_group: usize,
@@ -201,12 +230,14 @@ impl Kanva {
         Some(first..last + 1)
     }
 
-    /// Return a cursor for setting per-path field overrides at `path_idx`.
+    /// Return a cursor for setting per-path field overrides at
+    /// `path_idx`.
     pub fn mod_path(&mut self, path_idx: usize) -> PathModEntry<'_> {
         PathModEntry::new(&mut self.path_mods, path_idx)
     }
 
-    /// Return a cursor for setting per-group field overrides at `group_idx`.
+    /// Return a cursor for setting per-group field overrides at
+    /// `group_idx`.
     pub fn mod_group(
         &mut self,
         group_idx: usize,
@@ -224,8 +255,8 @@ impl Kanva {
 
     /// Render into any [`PaintSink`].
     ///
-    /// Group transforms are accumulated and multiplied by each path's stored
-    /// world transform at draw time.
+    /// Group transforms are accumulated and multiplied by each path's
+    /// stored world transform at draw time.
     pub fn render(&self, sink: &mut impl PaintSink) {
         let mut group_tf_stack = vec![Affine::IDENTITY];
 
@@ -316,34 +347,48 @@ impl Kanva {
                     } else {
                         path.fill.map(|i| &self.fills[i])
                     };
-                    if let Some(fill) = fill {
-                        sink.fill(FillRef {
-                            transform: eff_tf,
-                            fill_rule: fill.rule,
-                            brush: (&fill.brush).into(),
-                            brush_transform: fill.brush_transform,
-                            shape: GeometryRef::Path(data),
-                            composite: fill.composite,
-                        });
-                    }
-
-                    let stroke_style = if let Some(stroke_mod) =
+                    let stroke = if let Some(stroke_mod) =
                         self.path_mods.stroke.get(&idx)
                     {
                         stroke_mod.as_ref()
                     } else {
                         path.stroke.map(|i| &self.strokes[i])
                     };
-                    if let Some(stroke_style) = stroke_style {
-                        sink.stroke(StrokeRef {
-                            transform: eff_tf,
-                            stroke: &stroke_style.stroke,
-                            brush: (&stroke_style.brush).into(),
-                            brush_transform: stroke_style
-                                .brush_transform,
-                            shape: GeometryRef::Path(data),
-                            composite: stroke_style.composite,
-                        });
+
+                    let emit_fill = |sink: &mut dyn PaintSink| {
+                        if let Some(fill) = fill {
+                            sink.fill(FillRef {
+                                transform: eff_tf,
+                                fill_rule: fill.rule,
+                                brush: (&fill.brush).into(),
+                                brush_transform: fill.brush_transform,
+                                shape: GeometryRef::Path(data),
+                                composite: fill.composite,
+                            });
+                        }
+                    };
+                    let emit_stroke = |sink: &mut dyn PaintSink| {
+                        if let Some(stroke) = stroke {
+                            sink.stroke(StrokeRef {
+                                transform: eff_tf,
+                                stroke: &stroke.stroke,
+                                brush: (&stroke.brush).into(),
+                                brush_transform: stroke
+                                    .brush_transform,
+                                shape: GeometryRef::Path(data),
+                                composite: stroke.composite,
+                            });
+                        }
+                    };
+                    match path.paint_order {
+                        PaintOrder::FillStroke => {
+                            emit_fill(sink);
+                            emit_stroke(sink);
+                        }
+                        PaintOrder::StrokeFill => {
+                            emit_stroke(sink);
+                            emit_fill(sink);
+                        }
                     }
 
                     if alpha.is_some() {
@@ -357,12 +402,14 @@ impl Kanva {
 
 #[cfg(test)]
 mod tests {
-    use super::*;
-    use crate::builder::KanvaBuilder;
-    use crate::sink::KanvaSink;
     use imaging::kurbo::{Affine, BezPath, Stroke};
     use imaging::peniko::{Brush, Color, Fill};
-    use imaging::record::{Command as RecCmd, Draw, Scene};
+    use imaging::record::{Command, Draw, Scene};
+
+    use crate::builder::KanvaBuilder;
+    use crate::sink::KanvaSink;
+
+    use super::*;
 
     fn build_fill(brush: &Brush, transform: Affine) -> Kanva {
         let mut b = KanvaBuilder::new();
@@ -376,6 +423,7 @@ mod tests {
                 ..Default::default()
             }),
             None,
+            PaintOrder::default(),
         );
         b.build()
     }
@@ -392,8 +440,62 @@ mod tests {
                 brush: Brush::default(),
                 ..Default::default()
             }),
+            PaintOrder::default(),
         );
         b.build()
+    }
+
+    #[test]
+    fn combined_fill_stroke_emits_fill_then_stroke() {
+        let mut b = KanvaBuilder::new();
+        KanvaSink::draw_path(
+            &mut b,
+            BezPath::new(),
+            Affine::IDENTITY,
+            Some(KanvaFill {
+                rule: Fill::NonZero,
+                brush: Brush::default(),
+                ..Default::default()
+            }),
+            Some(KanvaStroke {
+                stroke: Stroke::default(),
+                brush: Brush::default(),
+                ..Default::default()
+            }),
+            PaintOrder::default(),
+        );
+        let kanva = b.build();
+        assert_eq!(
+            kanva.paths().len(),
+            1,
+            "one KanvaPath for both fill+stroke"
+        );
+        let mut scene = Scene::new();
+        kanva.render(&mut scene);
+        let cmds = scene.commands();
+        assert_eq!(cmds.len(), 2, "two draw ops from one path");
+        assert!(
+            matches!(cmds[0], Command::Draw(_)),
+            "first op is Draw"
+        );
+        assert!(
+            matches!(cmds[1], Command::Draw(_)),
+            "second op is Draw"
+        );
+        let Command::Draw(fill_id) = cmds[0] else {
+            unreachable!()
+        };
+        let Command::Draw(stroke_id) = cmds[1] else {
+            unreachable!()
+        };
+        assert!(
+            matches!(scene.draw_op(fill_id), Draw::Fill { .. }),
+            "first draw is Fill"
+        );
+        assert!(
+            matches!(scene.draw_op(stroke_id), Draw::Stroke { .. }),
+            "second draw is Stroke"
+        );
     }
 
     #[test]
@@ -403,8 +505,8 @@ mod tests {
         kanva.render(&mut scene);
         let cmds = scene.commands();
         assert_eq!(cmds.len(), 1);
-        let RecCmd::Draw(id) = cmds[0] else {
-            panic!("expected Draw")
+        let Command::Draw(id) = cmds[0] else {
+            panic!("expected Command::Draw, got {:?}", cmds[0])
         };
         assert!(matches!(scene.draw_op(id), Draw::Fill { .. }));
     }
@@ -416,8 +518,8 @@ mod tests {
         kanva.render(&mut scene);
         let cmds = scene.commands();
         assert_eq!(cmds.len(), 1);
-        let RecCmd::Draw(id) = cmds[0] else {
-            panic!("expected Draw")
+        let Command::Draw(id) = cmds[0] else {
+            panic!("expected Command::Draw, got {:?}", cmds[0])
         };
         assert!(matches!(scene.draw_op(id), Draw::Stroke { .. }));
     }
@@ -437,6 +539,7 @@ mod tests {
                 ..Default::default()
             }),
             None,
+            PaintOrder::default(),
         );
         KanvaSink::pop_group(&mut b);
         let kanva = b.build();
@@ -444,9 +547,9 @@ mod tests {
         kanva.render(&mut scene);
         let cmds = scene.commands();
         assert_eq!(cmds.len(), 3);
-        assert!(matches!(cmds[0], RecCmd::PushGroup(_)));
-        assert!(matches!(cmds[1], RecCmd::Draw(_)));
-        assert!(matches!(cmds[2], RecCmd::PopGroup));
+        assert!(matches!(cmds[0], Command::PushGroup(_)));
+        assert!(matches!(cmds[1], Command::Draw(_)));
+        assert!(matches!(cmds[2], Command::PopGroup));
     }
 
     #[test]
@@ -467,6 +570,7 @@ mod tests {
                 ..Default::default()
             }),
             None,
+            PaintOrder::default(),
         );
         KanvaSink::pop_group(&mut b);
         let mut kanva = b.build();
@@ -475,11 +579,14 @@ mod tests {
         let mut scene = Scene::new();
         kanva.render(&mut scene);
         let cmds = scene.commands();
-        let RecCmd::Draw(id) = cmds[1] else {
-            panic!("expected Draw")
+        let Command::Draw(id) = cmds[1] else {
+            panic!(
+                "expected Command::Draw at index 1, got {:?}",
+                cmds[1]
+            )
         };
         let Draw::Fill { transform, .. } = scene.draw_op(id) else {
-            panic!()
+            panic!("expected Draw::Fill, got {:?}", scene.draw_op(id))
         };
         assert_eq!(*transform, scale * path_tf);
     }
@@ -498,11 +605,14 @@ mod tests {
         );
         let mut scene = Scene::new();
         kanva.render(&mut scene);
-        let RecCmd::Draw(id) = scene.commands()[0] else {
-            panic!()
+        let Command::Draw(id) = scene.commands()[0] else {
+            panic!(
+                "expected Command::Draw, got {:?}",
+                scene.commands()[0]
+            )
         };
         let Draw::Fill { brush, .. } = scene.draw_op(id) else {
-            panic!()
+            panic!("expected Draw::Fill, got {:?}", scene.draw_op(id))
         };
         assert_eq!(*brush, override_brush);
     }
@@ -516,12 +626,12 @@ mod tests {
         kanva.render(&mut scene);
         let cmds = scene.commands();
         assert_eq!(cmds.len(), 3);
-        let RecCmd::PushGroup(gid) = cmds[0] else {
-            panic!("expected PushGroup")
+        let Command::PushGroup(gid) = cmds[0] else {
+            panic!("expected Command::PushGroup, got {:?}", cmds[0])
         };
         assert_eq!(scene.group(gid).composite.alpha, 0.5);
-        assert!(matches!(cmds[1], RecCmd::Draw(_)));
-        assert!(matches!(cmds[2], RecCmd::PopGroup));
+        assert!(matches!(cmds[1], Command::Draw(_)));
+        assert!(matches!(cmds[2], Command::PopGroup));
     }
 
     #[test]
@@ -532,11 +642,14 @@ mod tests {
         kanva.path_mods.transform(0, override_tf);
         let mut scene = Scene::new();
         kanva.render(&mut scene);
-        let RecCmd::Draw(id) = scene.commands()[0] else {
-            panic!()
+        let Command::Draw(id) = scene.commands()[0] else {
+            panic!(
+                "expected Command::Draw, got {:?}",
+                scene.commands()[0]
+            )
         };
         let Draw::Fill { transform, .. } = scene.draw_op(id) else {
-            panic!()
+            panic!("expected Draw::Fill, got {:?}", scene.draw_op(id))
         };
         assert_eq!(*transform, override_tf);
     }
@@ -560,6 +673,7 @@ mod tests {
                 ..Default::default()
             }),
             None,
+            PaintOrder::default(),
         );
         KanvaSink::pop_group(&mut b);
         let mut kanva = b.build();
@@ -568,11 +682,14 @@ mod tests {
 
         let mut scene = Scene::new();
         kanva.render(&mut scene);
-        let RecCmd::Draw(id) = scene.commands()[1] else {
-            panic!()
+        let Command::Draw(id) = scene.commands()[1] else {
+            panic!(
+                "expected Command::Draw at index 1, got {:?}",
+                scene.commands()[1]
+            )
         };
         let Draw::Fill { transform, .. } = scene.draw_op(id) else {
-            panic!()
+            panic!("expected Draw::Fill, got {:?}", scene.draw_op(id))
         };
         assert_eq!(*transform, override_tf * path_tf);
     }
@@ -593,14 +710,18 @@ mod tests {
                 ..Default::default()
             }),
             None,
+            PaintOrder::default(),
         );
         KanvaSink::pop_group(&mut b);
         let mut kanva = b.build();
         kanva.group_mods.composite(0, composite);
         let mut scene = Scene::new();
         kanva.render(&mut scene);
-        let RecCmd::PushGroup(gid) = scene.commands()[0] else {
-            panic!()
+        let Command::PushGroup(gid) = scene.commands()[0] else {
+            panic!(
+                "expected Command::PushGroup, got {:?}",
+                scene.commands()[0]
+            )
         };
         assert_eq!(scene.group(gid).composite, composite);
     }
@@ -621,14 +742,17 @@ mod tests {
         // First render: override active.
         let mut scene = Scene::new();
         kanva.render(&mut scene);
-        let RecCmd::Draw(id) = scene.commands()[0] else {
-            panic!()
+        let Command::Draw(id) = scene.commands()[0] else {
+            panic!(
+                "expected Command::Draw, got {:?}",
+                scene.commands()[0]
+            )
         };
         let Draw::Fill {
             brush: first_brush, ..
         } = scene.draw_op(id)
         else {
-            panic!()
+            panic!("expected Draw::Fill, got {:?}", scene.draw_op(id))
         };
         assert_eq!(*first_brush, Brush::Solid(Color::WHITE));
 
@@ -637,15 +761,18 @@ mod tests {
         // Second render: original restored.
         let mut scene = Scene::new();
         kanva.render(&mut scene);
-        let RecCmd::Draw(id) = scene.commands()[0] else {
-            panic!()
+        let Command::Draw(id) = scene.commands()[0] else {
+            panic!(
+                "expected Command::Draw, got {:?}",
+                scene.commands()[0]
+            )
         };
         let Draw::Fill {
             brush: second_brush,
             ..
         } = scene.draw_op(id)
         else {
-            panic!()
+            panic!("expected Draw::Fill, got {:?}", scene.draw_op(id))
         };
         assert_eq!(*second_brush, original);
     }
