+++ crates/kanva/src/node.rs
@@ -7,11 +7,13 @@ pub(crate) const PATH_TOLERANCE: f64 = 0.1;
 
 /// A group in the [`Kanva`][crate::Kanva] scene graph.
 ///
-/// Groups carry a transform that is accumulated onto child path world transforms
-/// at render time. They may also carry a clip shape and a composite mode.
+/// Groups carry a transform that is accumulated onto child path world
+/// transforms at render time. They may also carry a clip shape and a
+/// composite mode.
 #[derive(Default, Debug, Clone)]
 pub struct Group {
-    /// Transform accumulated with parent group transforms at render time.
+    /// Transform accumulated with parent group transforms at render
+    /// time.
     pub transform: Affine,
     pub clip: Option<KanvaClip>,
     pub composite: Composite,
@@ -52,22 +54,39 @@ impl KanvaClip {
     }
 }
 
-/// A stored path with its world transform and optional fill/stroke indices.
+/// Controls whether fill or stroke is drawn first for a
+/// [`KanvaPath`].
+#[derive(Default, Debug, Clone, Copy, PartialEq, Eq)]
+pub enum PaintOrder {
+    /// Fill is drawn before stroke (default, matches SVG
+    /// `fill-and-stroke`).
+    #[default]
+    FillStroke,
+    /// Stroke is drawn before fill (matches SVG `stroke-and-fill`).
+    StrokeFill,
+}
+
+/// A stored path with its world transform and optional fill/stroke
+/// indices.
 #[derive(Default, Debug, Clone)]
 pub struct KanvaPath {
     pub path: BezPath,
     /// Full world transform as received from [`imaging`].
     pub transform: Affine,
-    /// Index into the fills buffer; retrieve via [`crate::Kanva::get_fill`].
+    /// Index into the fills buffer; retrieve via
+    /// [`crate::Kanva::get_fill`].
     pub fill: Option<usize>,
-    /// Index into the strokes buffer; retrieve via [`crate::Kanva::get_stroke`].
+    /// Index into the strokes buffer; retrieve via
+    /// [`crate::Kanva::get_stroke`].
     pub stroke: Option<usize>,
+    pub paint_order: PaintOrder,
 }
 
 /// A draw command in the [`Kanva`][crate::Kanva] command buffer.
 #[derive(Debug, Clone, Copy)]
 pub enum Command {
-    /// Push a group onto the render stack (index into the groups vec).
+    /// Push a group onto the render stack (index into the groups
+    /// vec).
     PushGroup(usize),
     /// Pop the current group from the render stack.
     PopGroup,
@@ -75,10 +94,12 @@ pub enum Command {
     DrawPath(usize),
 }
 
-/// Command index range for a group (parallel to [`crate::Kanva`]'s `groups` vec).
+/// Command index range for a group (parallel to [`crate::Kanva`]'s
+/// `groups` vec).
 ///
-/// The inner commands are at `start + 1..end`: `start` is the `PushGroup`
-/// and `end` is the `PopGroup`, both excluded when iterating contents.
+/// The inner commands are at `start + 1..end`: `start` is the
+/// `PushGroup` and `end` is the `PopGroup`, both excluded when
+/// iterating contents.
 #[derive(Debug, Clone, Copy)]
 pub struct GroupRange {
     /// Index of the `PushGroup` command for this group.
